+++ src/ts/options/binding.ts
@@ -36,14 +36,17 @@ export namespace Binding {
             options.json = Default.getString( options.json, propertyDefaults.json );
             options.templateData = Default.getObject( options.templateData, propertyDefaults.templateData );
             options.removeOriginalAttributes = Default.getBoolean( options.removeOriginalAttributes, propertyDefaults.removeOriginalAttributes );
+            options.removeOriginalDataAttributes = Default.getBoolean( options.removeOriginalDataAttributes, propertyDefaults.removeOriginalDataAttributes );
             options.clearOriginalHTML = Default.getBoolean( options.clearOriginalHTML, propertyDefaults.clearOriginalHTML );
             options.addCssToHead = Default.getBoolean( options.addCssToHead, propertyDefaults.addCssToHead );
             options.clearCssFromHead = Default.getBoolean( options.clearCssFromHead, propertyDefaults.clearCssFromHead );
             options.logTemplateDataWarnings = Default.getBoolean( options.logTemplateDataWarnings, propertyDefaults.logTemplateDataWarnings );
             options.addAttributes = Default.getBoolean( options.addAttributes, propertyDefaults.addAttributes );
+            options.addDataAttributes = Default.getBoolean( options.addDataAttributes, propertyDefaults.addDataAttributes );
             options.addCssProperties = Default.getBoolean( options.addCssProperties, propertyDefaults.addCssProperties );
             options.addText = Default.getBoolean( options.addText, propertyDefaults.addText );
             options.addChildren = Default.getBoolean( options.addChildren, propertyDefaults.addChildren );
+            options.insertBefore = Default.getBoolean( options.insertBefore, propertyDefaults.insertBefore );
 
             options = getCustomTriggers( options );